qrcode/requirements.txt
pillow
numpy
imageio
